third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-outline-offset.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>

description("Test to make sure outline-offset is found in the computed style.")

var testContainer = document.createElement("div");
document.body.appendChild(testContainer);

testContainer.innerHTML = '<div id="test" style="outline: 2px solid black">hello</div>';

e = document.getElementById('test');
e.contentEditable = true;

e.style.outlineOffset = "2px";
computedStyle = window.getComputedStyle(e);
shouldBe("computedStyle.getPropertyValue('outline-offset')", "'2px'");

document.body.removeChild(testContainer);

</script>
</body>
</html>
